+++ src/common/components/Layouts/Home/HomeNavbar/Menu/index.tsx
@@ -6,7 +6,7 @@ import { APP_NAVBAR_CONSTANT } from 'Models/menu/constant';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React from 'react';
-import pathJoin from 'Utils/path-join';
+import pathJoin from 'src/common/utils/path-join';
 
 const Menu: React.FC = () => {
   const router = useRouter();
