+++ src/common/components/Layouts/Home/HomeNavbar/SideMenu/SideMenu.spec.tsx
