+++ src/common/models/oauth/type.ts
