+++ src/common/tests/mock/fileMock.js
