+++ src/pages/_app.tsx
@@ -4,8 +4,8 @@ import { AppProps } from 'next/dist/next-server/lib/router/router';
 import { useRouter } from 'next/router';
 import React from 'react';
 import { Provider } from 'react-redux';
-import { store } from 'Stores/stores';
-import { GlobalStyle } from 'Styles/Global.styled';
+import { store } from 'src/common/stores/stores';
+import { GlobalStyle } from 'src/common/styles/Global.styled';
 
 const EXCLUDE_NAVBAR: string[] = ['/login'];
