+++ src/pages/index.tsx
@@ -1,10 +1,10 @@
 import { Card, Col, Row } from 'antd';
 import Button from 'Components/Button';
-import HomeLayout from 'Components/Layouts/Home';
+import HomeLayout from 'src/common/tests/pages/node_modules/Components/Layouts/Home';
 import { APP_FEATURE_CONSTANT } from 'Models/featureCard/constant';
 import React from 'react';
 import styled, { CSSProperties } from 'styled-components';
-import IconSelector from 'Utils/icon-selector';
+import IconSelector from 'src/common/utils/icon-selector';
 
 const Icon = styled.div`
   font-size: 40px;
